super classed EventList to TimeSeries
