+++ threeML/utils/time_series/event_polynomial.py
